refactor: move timestamp index out of spentindex
