+++ src/Makefile.am
@@ -297,6 +297,7 @@ BITCOIN_CORE_H = \
   support/events.h \
   support/lockedpool.h \
   sync.h \
+  timestampindex.h \
   threadsafety.h \
   threadinterrupt.h \
   timedata.h \
